@@ -8,7 +8,7 @@ write_to = "src/_gettsim/_version.py"
 
 
 [tool.ruff]
-target-version = "py39"
+target-version = "py311"
 select = ["ALL"]
 fix = true
 extend-ignore = [
@@ -45,6 +45,8 @@ extend-ignore = [
     "DTZ002", # use of `datetime.datetime.today()` is not allowed
     "PT012", # `pytest.raises()` block should contain a single simple statement
     "PLR5501", # elif not supported by Jax converter
+    "TRY003", # Avoid specifying long messages outside the exception class
+    "FIX002", # Line contains TODO -- Use stuff from TD area.
 
     # Things we are not sure we want
     # ==============================
@@ -66,27 +68,21 @@ extend-ignore = [
 exclude = []
 
 [tool.ruff.per-file-ignores]
-"src/_gettsim_tests/test_docs.py" = ["TRY"]
 "src/_gettsim_tests/test_rounding.py" = ["PT019"]
-"src/_gettsim/aggregation_numpy.py" = ["TRY"]
 "src/_gettsim/benefits/elterngeld.py" = ["E501"]
 "src/_gettsim/benefits/kinderzuschl.py" = ["ARG001"]
 "src/_gettsim/benefits/unterhalt.py" = ["E501"]
 "src/_gettsim/benefits/wohngeld.py" = ["ARG001"]
-"src/_gettsim/config.py" = ["TRY"]
 "src/_gettsim/functions.py" = ["F401"]
-"src/_gettsim/functions_loader.py" = ["TRY"]
-"src/_gettsim/gettsim_typing.py" = ["PGH", "PLR", "TRY", "SIM114"]
-"src/_gettsim/interface.py" = ["TRY"]
-"src/_gettsim/piecewise_functions.py" = ["PLR", "TRY"]
-"src/_gettsim/policy_environment.py" = ["S506", "PLR", "TRY"]
+"src/_gettsim/gettsim_typing.py" = ["PGH", "PLR", "SIM114"]
+"src/_gettsim/piecewise_functions.py" = ["PLR"]
+"src/_gettsim/policy_environment.py" = ["S506", "PLR"]
 "src/_gettsim/rente.py" = ["N806", "ARG001"]
 "src/_gettsim/social_insurance_contributions/ges_krankenv.py" = ["SIM401"]
-"src/_gettsim/synthetic.py" = ["TRY"]
 "src/_gettsim/tax_transfer.py" = ["ARG001"]
 "src/_gettsim/taxes/zu_versteuerndes_eink.py" = ["E501", "ARG001"]
 "src/_gettsim/tests/test_tax_transfer.py" = ["E501"]
-"src/_gettsim/visualization.py" = ["TRY", "PLR"]
+"src/_gettsim/visualization.py" = ["PLR"]
 
 [tool.ruff.pydocstyle]
 convention = "numpy"
@@ -131,6 +127,7 @@ filterwarnings = [
     "ignore:Repeated execution of the test suite",
     "ignore:Using or importing the ABCs from 'collections'",
     "ignore:'grouped_cumsum' is deprecated.",
+    "ignore::_gettsim.interface.FunctionsAndColumnsOverlapWarning"
 ]
 markers = [
     "wip: Tests that are work-in-progress.",
