@@ -146,6 +146,7 @@ TYPES_INPUT_VARIABLES = {
     "bruttokaltmiete_m_hh": float,
     "heizkosten_m_hh": float,
     "jahr_renteneintr": int,
+    "monat_renteneintr": int,
     "behinderungsgrad": int,
     "wohnfläche_hh": float,
     "m_elterngeld": int,
@@ -166,7 +167,7 @@ TYPES_INPUT_VARIABLES = {
     "m_mutterschutz": float,
     "m_arbeitsunfähig": float,
     "m_krank_ab_16_bis_24": float,
-    "m_arbeitslos": float,
+    "m_arbeitsl": float,
     "m_ausbild_suche": float,
     "m_schul_ausbild": float,
     "m_geringf_beschäft": float,
@@ -175,6 +176,8 @@ TYPES_INPUT_VARIABLES = {
     "m_kind_berücks_zeit": float,
     "m_pfleg_berücks_zeit": float,
     "y_pflichtbeitr_ab_40": float,
+    "pflichtbeitr_8_in_10": bool,
+    "arbeitsl_1y_past_585": bool,
     "anwartschaftszeit": bool,
     "arbeitssuchend": bool,
     "m_durchg_alg1_bezug": float,
