@@ -1,7 +1,47 @@
 from _gettsim.piecewise_functions import piecewise_polynomial
+from _gettsim.shared import dates_active
 
 
-def soli_st_y_tu(
+@dates_active(end="2008-12-31", change_name="soli_st_y_tu")
+def soli_st_y_tu_ohne_abgelt_st(
+    eink_st_mit_kinderfreib_y_tu: float,
+    anz_erwachsene_tu: int,
+    soli_st_params: dict,
+) -> float:
+    """Calculate the Solidarity Surcharge on tax unit level.
+
+    Solidaritätszuschlaggesetz (SolZG) in 1991 and 1992.
+    Solidaritätszuschlaggesetz 1995 (SolZG 1995) since 1995.
+
+    The Solidarity Surcharge is an additional tax on top of the income tax which
+    is the tax base. As opposed to the 'standard' income tax, child allowance is
+    always deducted for tax base calculation.
+
+    There is also Solidarity Surcharge on the Capital Income Tax, but always
+    with Solidarity Surcharge tax rate and no tax exempt level. §3 (3) S.2
+    SolzG 1995.
+
+    Parameters
+    ----------
+    eink_st_mit_kinderfreib_y_tu
+        See :func:`eink_st_mit_kinderfreib_y_tu`.
+    anz_erwachsene_tu
+        See :func:`anz_erwachsene_tu`.
+    soli_st_params
+        See params documentation :ref:`soli_st_params <soli_st_params>`.
+
+    Returns
+    -------
+
+    """
+    eink_st_per_individual = eink_st_mit_kinderfreib_y_tu / anz_erwachsene_tu
+    out = anz_erwachsene_tu * _soli_st_tarif(eink_st_per_individual, soli_st_params)
+
+    return out
+
+
+@dates_active(start="2009-01-01", change_name="soli_st_y_tu")
+def soli_st_y_tu_mit_abgelt_st(
     eink_st_mit_kinderfreib_y_tu: float,
     anz_personen_tu: int,
     abgelt_st_y_tu: float,
