@@ -27,7 +27,6 @@ OUT_COLS = [
     "wohngeld_m_hh",
     "unterhaltsvors_m_hh",
 ]
-OVERRIDE_COLS = ["sum_ges_rente_priv_rente_m"]
 
 data = load_policy_test_data("full_taxes_and_transfers")
 
@@ -44,12 +43,16 @@ def test_full_taxes_and_transfers(
     policy_params, policy_functions = cached_set_up_policy_environment(
         date=test_data.date
     )
+
+    out = OUT_COLS.copy()
+    if test_data.date.year <= 2008:
+        out.remove("abgelt_st_y_tu")
+
     compute_taxes_and_transfers(
         data=df,
         params=policy_params,
         functions=policy_functions,
-        targets=OUT_COLS,
-        columns_overriding_functions=OVERRIDE_COLS,
+        targets=out,
     )
 
 
@@ -64,6 +67,10 @@ def test_data_types(
     imports = _convert_paths_to_import_strings(PATHS_TO_INTERNAL_FUNCTIONS)
     functions = _load_functions(imports)
 
+    out = OUT_COLS.copy()
+    if test_data.date.year <= 2008:
+        out.remove("abgelt_st_y_tu")
+
     # Load all time dependent functions
     for y in range(1990, 2023):
         year_functions = load_functions_for_date(datetime.date(year=y, month=1, day=1))
@@ -77,9 +84,8 @@ def test_data_types(
         data=df,
         params=policy_params,
         functions=policy_functions,
-        targets=OUT_COLS,
+        targets=out,
         debug=True,
-        columns_overriding_functions=OVERRIDE_COLS,
     )
     for column_name, series in result.items():
         if series.empty:
@@ -92,11 +98,12 @@ def test_data_types(
             elif column_name in year_functions:
                 internal_type = year_functions[column_name].__annotations__["return"]
             else:
-                # ToDo: Implement easy way to find out expected type of
-                # ToDo: aggregated functions
+                # TODO (@hmgaudecker): Implement easy way to find out expected type of
+                #     aggregated functions
+                # https://github.com/iza-institute-of-labor-economics/gettsim/issues/604
                 if column_name.endswith(("_tu", "_bg", "_hh")):
                     internal_type = None
                 else:
-                    raise ValueError(f"Column name {column_name} unknown.")  # noqa: TRY
+                    raise ValueError(f"Column name {column_name} unknown.")
             if internal_type:
                 assert check_series_has_expected_type(series, internal_type)
